+++ windows/src/developer/TIKE/xml/layoutbuilder/physical-keyboard-template.js
@@ -1,0 +1,83 @@
+  "row": [
+     {
+      "id": 1,
+      "key": [
+        {"id": "K_1"},
+        {"id": "K_2"},
+        {"id": "K_3"},
+        {"id": "K_4"},
+        {"id": "K_5"},
+        {"id": "K_6"},
+        {"id": "K_7"},
+        {"id": "K_8"},
+        {"id": "K_9"},
+        {"id": "K_0"},
+        {"id": "K_HYPHEN"},
+        {"id": "K_EQUAL"},
+        {"id": "K_BKSP","text": "*BkSp*","width": "100","sp": "1"}
+      ]
+    },         
+    {
+      "id": 2,
+      "key": [
+        {"id": "K_Q","pad":75},
+        {"id": "K_W"},
+        {"id": "K_E"},
+        {"id": "K_R"},
+        {"id": "K_T"},
+        {"id": "K_Y"},
+        {"id": "K_U"},
+        {"id": "K_I"},
+        {"id": "K_O"},
+        {"id": "K_P"},
+        {"id": "K_LBRKT"},
+        {"id": "K_RBRKT"},
+        {"sp":"10","width":"10"}              
+      ]
+    },
+    {
+      "id": 3,
+      "key": [
+        {"id": "K_BKQUOTE"},
+        {"id": "K_A"},
+        {"id": "K_S"},
+        {"id": "K_D"},
+        {"id": "K_F"},
+        {"id": "K_G"},
+        {"id": "K_H"},
+        {"id": "K_J"},
+        {"id": "K_K"},
+        {"id": "K_L"},
+        {"id": "K_COLON"},
+        {"id": "K_QUOTE"},
+        {"id": "K_BKSLASH"}
+      ]
+    },
+    {
+      "id": 4,
+      "key": [
+        {"id": "K_SHIFT","text": "*Shift*","width": "160","sp": "1"},
+        {"id": "K_oE2"},
+        {"id": "K_Z"},
+        {"id": "K_X"},
+        {"id": "K_C"},
+        {"id": "K_V"},
+        {"id": "K_B"},
+        {"id": "K_N"},
+        {"id": "K_M"},
+        {"id": "K_COMMA"},
+        {"id": "K_PERIOD"},
+        {"id": "K_SLASH"},
+        {"sp":"10","width":"10"}              
+      ]
+    },
+    {
+      "id": 5,
+      "key": [
+        {"id": "K_LCONTROL","text": "*Ctrl*","width": "130","sp": "1"},
+        {"id": "K_LOPT","text": "*Menu*","width": "140","sp": "1"},
+        {"id": "K_SPACE","text": "","width": "930","sp": "0"},
+        {"id": "K_ENTER","text": "*Enter*","width":"145","sp": "1"}
+      ]
+    }
+  ]