+++ web/history.md
@@ -10,8 +10,9 @@
 * Refactored code paths relating to KMW's embedding in the iOS and Android apps.  (#211)
 * Added support for L/R Alt and Ctrl modifiers for keyboards. (#9) (#52)
 * Added support for use of the Caps Lock state within keyboards if specified by a keyboard designer.
-* Fixed next-layer processing (#116)
 * Fixed keyboard loading upon initialization, redundant cloud requests for keyboards.  (#103)
+* Fixed next-layer processing (#116) (#358)
+* Fixed auto-attaching mode bug. (#352)
 
 ## 2017-07-10 2.0.473 stable
 * 2.0 stable release build.